requirements.txt
asgiref==3.8.1
boto3==1.34.127
botocore==1.34.127
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-admin-sortable2==2.2.1
django-appconf==1.0.6
django-embed-video==1.4.10
django-environ==0.11.2
django-heroku==0.3.1
django-imagekit==5.0.0
django-optimized-image==0.3.0
django-s3-storage==0.15.0
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
packaging==24.1
pilkit==3.0
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tinify==1.6.0
typing_extensions==4.12.2
urllib3==2.2.1
whitenoise==6.6.0
